+++ shimmingtoolbox/optimizer/lsq_optimizer.py
@@ -6,21 +6,27 @@ import scipy.optimize as opt
 from typing import List
 import warnings
 
-from shimmingtoolbox.optimizer.basic_optimizer import Optimizer
+from shimmingtoolbox.optimizer.optimizer_utils import OptimizerUtils
 from shimmingtoolbox.pmu import PmuResp
 from shimmingtoolbox.coils.coil import Coil
 
 ListCoil = List[Coil]
 allowed_opt_criteria = ['mse', 'mae', 'std']
 
+import logging
 
-class LsqOptimizer(Optimizer):
+
+logger = logging.getLogger(__name__)
+
+
+class LsqOptimizer(OptimizerUtils):
     """ Optimizer object that stores coil profiles and optimizes an unshimmed volume given a mask.
         Use optimize(args) to optimize a given mask. The algorithm uses a least squares solver to find the best shim.
         It supports bounds for each channel as well as a bound for the absolute sum of the channels.
     """
 
-    def __init__(self, coils: ListCoil, unshimmed, affine, opt_criteria='mse', reg_factor=0):
+    def __init__(self, coils: ListCoil, unshimmed, affine, opt_criteria='mse', initial_guess_method='mean',
+                 reg_factor=0):
         """
         Initializes coils according to input list of Coil
 
@@ -34,12 +40,6 @@ class LsqOptimizer(Optimizer):
                                 penalize higher current values while a lower factor will lower the effect of the
                                 regularization. A negative value will favour high currents (not preferred).
         """
-        super().__init__(coils, unshimmed, affine)
-        self._initial_guess_method = 'mean'
-        self.initial_coefs = None
-        self.reg_factor = reg_factor
-        self.reg_factor_channel = np.array([max(np.abs(bound)) for bound in self.merged_bounds])
-        self.reg_vector = self.reg_factor / (len(self.reg_factor_channel) * self.reg_factor_channel)
         lsq_residual_dict = {
             allowed_opt_criteria[0]: self._residuals_mse,
             allowed_opt_criteria[1]: self._residuals_mae,
@@ -57,24 +57,7 @@ class LsqOptimizer(Optimizer):
             self.opt_criteria = opt_criteria
         else:
             raise ValueError("Optimization criteria not supported")
-
-    @property
-    def initial_guess_method(self):
-        return self._initial_guess_method
-
-    @initial_guess_method.setter
-    def initial_guess_method(self, method, coefs=None):
-        allowed_methods = ['mean', 'zeros', 'set']
-        if method not in allowed_methods:
-            raise ValueError(f"Initial_guess_method not supported. Supported methods are: {allowed_methods}")
-
-        if method == 'set':
-            if coefs is not None:
-                self.initial_coefs = coefs
-            else:
-                raise ValueError(f"There are no coefficients to set")
-
-        self._initial_guess_method = method
+        super().__init__(coils, unshimmed, affine, initial_guess_method, reg_factor)
 
     def _residuals_mae(self, coef, unshimmed_vec, coil_mat, factor):
         """ Objective function to minimize the mean absolute error (MAE)
@@ -213,99 +196,21 @@ class LsqOptimizer(Optimizer):
                                        constraints=tuple(scipy_constraints),
                                        jac=self._jacobian_func,
                                        options={'maxiter': 1000})
+
         else:
-            currents_sp = opt.minimize(self._criteria_func, currents_0,
-                                       args=(unshimmed_vec, coil_mat, factor),
-                                       method='SLSQP',
-                                       bounds=self.merged_bounds,
-                                       constraints=tuple(scipy_constraints),
-                                       jac=self._jacobian_func,
-                                       options={'maxiter': 1000})
+
+            currents_sp = opt.minimize(self._initial_guess_mse, currents_0,
+                                        args=(unshimmed_vec, coil_mat, factor),
+                                        method='SLSQP',
+                                        bounds=self.merged_bounds,
+                                        constraints=tuple(scipy_constraints),
+                                        options={'maxiter': 1000})
 
         return currents_sp
 
-    def get_initial_guess(self):
-        """ Calculates the initial guess according to the `self.initial_guess_method`
+    def _get_currents(self, unshimmed_vec, coil_mat, currents_0):
 
-        Returns:
-            np.ndarray: 1d array (n_channels) containing the initial guess for the optimization
-        """
-
-        allowed_guess_method = {
-            'mean': self._initial_guess_mean_bounds,
-            'zeros': self._initial_guess_zeros,
-            'set': self._initial_guess_set
-        }
-
-        initial_guess = allowed_guess_method[self.initial_guess_method]()
-
-        return initial_guess
-
-    def _initial_guess_set(self):
-        return self.initial_coefs
-
-    def _initial_guess_mean_bounds(self):
-        """
-        Calculates the initial guess from the bounds, sets it to the mean of the bounds
-
-        Returns:
-            np.ndarray: 1d array (n_channels) of coefficient representing the initial guess
-
-        """
-        current_0 = []
-        for bounds in self.merged_bounds:
-            avg = np.mean(bounds)
-
-            if np.isnan(avg):
-                current_0.append(0)
-            else:
-                current_0.append(avg)
-
-        return np.array(current_0)
-
-    def _initial_guess_zeros(self):
-        """
-        Return a numpy array with zeros.
-
-        Returns:
-            np.ndarray: 1d array (n_channels) of coefficient representing the initial guess
-
-        """
-        current_0 = np.zeros(len(self.merged_bounds))
-
-        return current_0
-
-    def optimize(self, mask):
-        """
-        Optimize unshimmed volume by varying current to each channel
-
-        Args:
-            mask (numpy.ndarray): 3D integer mask used for the optimizer (only consider voxels with non-zero values).
-
-        Returns:
-            numpy.ndarray: Coefficients corresponding to the coil profiles that minimize the objective function.
-                           The shape of the array returned has shape corresponding to the total number of channels
-        """
-
-        # Check for sizing errors
-        self._check_sizing(mask)
-        # Define coil profiles
-        n_channels = self.merged_coils.shape[3]
-        mask_vec = mask.reshape((-1,))
-        # # Reshape coil profile: X, Y, Z, N --> [mask.shape], N
-        # #   --> N, [mask.shape] --> N, mask.size --> mask.size, N --> masked points, N
-        merged_coils_reshaped = np.reshape(np.transpose(self.merged_coils, axes=(3, 0, 1, 2)),
-                                           (n_channels, -1))
-        masked_points_indices = np.where(mask_vec != 0)
-        coil_mat = merged_coils_reshaped[:, masked_points_indices[0]].T  # masked points x N
-        unshimmed_vec = np.reshape(self.unshimmed, (-1,))[masked_points_indices[0]]  # mV'
         scipy_constraints = self._define_scipy_constraints()
-
-        # Set up output currents
-        currents_0 = self.get_initial_guess()
-        # If what to shim is already 0s
-        if np.all(unshimmed_vec == 0):
-            return np.zeros(np.shape(currents_0))
         # Optimize
         # When clipping to bounds, scipy raises a warning. Since this can be frequent for our purposes, we ignore that
         # warning
@@ -327,7 +232,6 @@ class LsqOptimizer(Optimizer):
                                                        factor=1)
             currents_sp = self._scipy_minimize(currents_0, unshimmed_vec, coil_mat, scipy_constraints,
                                                factor=stability_factor)
-
         if not currents_sp.success:
             raise RuntimeError(f"Optimization failed due to: {currents_sp.message}")
 
@@ -358,10 +262,9 @@ class PmuLsqOptimizer(LsqOptimizer):
             pmu (PmuResp): PmuResp object containing the respiratory trace information.
         """
 
-        super().__init__(coils, unshimmed, affine, opt_criteria, reg_factor=reg_factor)
+        super().__init__(coils, unshimmed, affine, opt_criteria, initial_guess_method='zeros', reg_factor=reg_factor )
         self.pressure_min = pmu.min
         self.pressure_max = pmu.max
-        self.initial_guess_method = 'zeros'
 
     def _define_scipy_constraints(self):
         """Redefined from super() to include more constraints"""
@@ -463,6 +366,8 @@ class PmuLsqOptimizer(LsqOptimizer):
                                        constraints=tuple(scipy_constraints),
                                        jac=self._jacobian_func,
                                        options={'maxiter': 500})
+
+
         else:
             currents_sp = opt.minimize(self._criteria_func, currents_0,
                                        args=(unshimmed_vec, coil_mat, factor),
