+++ shimmingtoolbox/shim/sequencer.py
@@ -1266,17 +1266,17 @@ class RealTimeSequencer(Sequencer):
         shim_trace_static_riro = np.array(shim_trace_static_riro).reshape(n_shim, nt)
         shim_trace_riro = np.array(shim_trace_riro).reshape(n_shim, nt)
         unshimmed_trace = np.array(unshimmed_trace).reshape(n_shim, nt)
-            
-        if self.path_output is not None:    
+
+        if self.path_output is not None:
             # Plot before vs after shimming averaged on time
             shimmed_mask_avg = np.zeros(mask_full_binary.shape)
             np.divide(np.sum(np.mean(masked_shim_static_riro, axis=3), axis=3), np.sum(mask_fmap_cs, axis=3), where=mask_full_binary.astype(bool), out=shimmed_mask_avg)
             self.plot_full_mask(np.mean(unshimmed, axis=3), shimmed_mask_avg, mask_full_binary)
-            
+
             # Plot STD over time before and after shimming
             self.plot_full_time_std(unshimmed, masked_shim_static_riro, mask_fmap_cs, mask_full_binary)
-        
-        if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:            
+
+        if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:
             # plot results
             i_slice = 0
             i_shim = 0
@@ -1493,7 +1493,7 @@ class RealTimeSequencer(Sequencer):
                      f"\nstatic_shim_static_riro: {static_shim_static_riro}"
                      f"\nstatic_shim_riro: {static_shim_riro}"
                      f"\nstatic_unshimmed: {static_unshimmed}")
-    
+
     def plot_full_mask(self, unshimmed, shimmed_masked, mask):
         """
         Plot and save the static full mask
@@ -1504,11 +1504,11 @@ class RealTimeSequencer(Sequencer):
             mask (np.ndarray): Binary mask in the fieldmap space
         """
         # Plot
-     
+
         mt_unshimmed = montage(unshimmed)
         mt_unshimmed_masked = montage(unshimmed * mask)
         mt_shimmed_masked = montage(shimmed_masked)
-        
+
         metric_unshimmed_std = calculate_metric_within_mask(unshimmed, mask, metric='std')
         metric_shimmed_std = calculate_metric_within_mask(shimmed_masked, mask, metric='std')
         metric_unshimmed_mean = calculate_metric_within_mask(unshimmed, mask, metric='mean')
@@ -1566,8 +1566,8 @@ class RealTimeSequencer(Sequencer):
         # Transform shimmed field map to shape (x, y, z, time)
         sum_mask_fmap_cs =  np.sum(mask_fmap_cs, axis=3)
         mask_extended = np.repeat(mask[..., np.newaxis], masked_shim_static_riro.shape[-2], axis=-1)
-        
-        # Transpose is used to cater to numpy division order 
+
+        # Transpose is used to cater to numpy division order
         # (3, 2, 4) / (3, 2) Does not work
         # (4, 2, 3) / (2, 3) Does work
         #* Using out parameter in np.divide() prevents inconsistent results
@@ -1576,12 +1576,12 @@ class RealTimeSequencer(Sequencer):
 
         std_shimmed_masked = np.std(shimmed_masked, axis=-1, dtype=np.float64)
         std_unshimmed = np.std(unshimmed, axis=-1, dtype=np.float64)
-        
-        ## Plot 
+
+        ## Plot
         mt_unshimmed = montage(np.mean(unshimmed, axis=-1))
         mt_unshimmed_masked = montage(std_unshimmed * mask)
         mt_shimmed_masked = montage(std_shimmed_masked)
-        
+
         metric_unshimmed_mean = calculate_metric_within_mask(std_unshimmed, mask, metric='mean')
         metric_shimmed_mean = calculate_metric_within_mask(std_shimmed_masked, mask, metric='mean')
